refator(backend-common): extract test utilities to plugin api
