+++ packages/backend-common/api-report-testUtils.md
@@ -3,24 +3,19 @@
 > Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
 
 ```ts
-// @public
-export function overridePackagePathResolution(
-  options: OverridePackagePathResolutionOptions,
-): PackagePathResolutionOverride;
+import { overridePackagePathResolution as overridePackagePathResolution_2 } from '@backstage/backend-plugin-api/testUtils';
+import { OverridePackagePathResolutionOptions as OverridePackagePathResolutionOptions_2 } from '@backstage/backend-plugin-api/testUtils';
+import { PackagePathResolutionOverride as PackagePathResolutionOverride_2 } from '@backstage/backend-plugin-api/testUtils';
 
-// @public (undocumented)
-export interface OverridePackagePathResolutionOptions {
-  packageName: string;
-  path?: string;
-  paths?: {
-    [path in string]: string | (() => string);
-  };
-}
+// @public @deprecated (undocumented)
+export const overridePackagePathResolution: typeof overridePackagePathResolution_2;
 
-// @public (undocumented)
-export interface PackagePathResolutionOverride {
-  restore(): void;
-}
+// @public @deprecated (undocumented)
+export type OverridePackagePathResolutionOptions =
+  OverridePackagePathResolutionOptions_2;
+
+// @public @deprecated (undocumented)
+export type PackagePathResolutionOverride = PackagePathResolutionOverride_2;
 
 // (No @packageDocumentation comment for this package)
 ```
