+++ packages/backend-plugin-api/api-report-testUtils.md
@@ -1,0 +1,26 @@
+## API Report File for "@backstage/backend-plugin-api"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+// @public
+export function overridePackagePathResolution(
+  options: OverridePackagePathResolutionOptions,
+): PackagePathResolutionOverride;
+
+// @public (undocumented)
+export interface OverridePackagePathResolutionOptions {
+  packageName: string;
+  path?: string;
+  paths?: {
+    [path in string]: string | (() => string);
+  };
+}
+
+// @public (undocumented)
+export interface PackagePathResolutionOverride {
+  restore(): void;
+}
+
+// (No @packageDocumentation comment for this package)
+```
